refactor: trpc get schedule handler
