+++ packages/trpc/server/routers/viewer/availability/schedule/update.handler.ts
@@ -1,3 +1,4 @@
+import { transformScheduleToAvailabilityForClient } from "@calcom/lib";
 import { getAvailabilityFromSchedule } from "@calcom/lib/availability";
 import { hasEditPermissionForUserID } from "@calcom/lib/hasEditPermissionForUser";
 import { prisma } from "@calcom/prisma";
@@ -5,7 +6,7 @@ import { prisma } from "@calcom/prisma";
 import { TRPCError } from "@trpc/server";
 
 import type { TrpcSessionUser } from "../../../../trpc";
-import { convertScheduleToAvailability, setupDefaultSchedule } from "../util";
+import { setupDefaultSchedule } from "../util";
 import type { TUpdateInputSchema } from "./update.schema";
 
 type UpdateOptions = {
@@ -116,7 +117,7 @@ export const updateHandler = async ({ input, ctx }: UpdateOptions) => {
     },
   });
 
-  const userAvailability = convertScheduleToAvailability(schedule);
+  const userAvailability = transformScheduleToAvailabilityForClient(schedule);
 
   return {
     schedule,
